Create beautiful_numbers_13
A task where you need to find all 13-length and 13-base numbers s.t. first 6 numbers sum is equal to the last 6 numbers sum. 
For example:  
Number 0055237050A00 - suits, because 0+0+5+5+2+3 = 0+5+0+A+0+0
Number 1234AB988BABA - doesn't suit , as 1+2+3+4+A+B != 8+8+B+A+B+A 

Though my programm works with any base and any length. It is very fun, and was fun to make.